+++ mobx/lib/mobx.dart
@@ -65,4 +65,5 @@ export 'package:mobx/src/core.dart'
 export 'package:mobx/src/core/atom_extensions.dart';
 
 /// The current version as per `pubspec.yaml`
-const version = '2.0.7+1';
+
+const version = '2.0.7+3';