+++ mobx_examples/lib/connectivity/connectivity_store.dart
@@ -3,6 +3,7 @@ import 'package:mobx/mobx.dart';
 
 part 'connectivity_store.g.dart';
 
+// ignore: library_private_types_in_public_api
 class ConnectivityStore = _ConnectivityStore with _$ConnectivityStore;
 
 abstract class _ConnectivityStore with Store {