+++ mobx_examples/lib/dice/dice_counter.dart
@@ -5,6 +5,7 @@ import 'package:mobx/mobx.dart';
 part 'dice_counter.g.dart';
 
 // This is the class used by rest of your codebase
+// ignore: library_private_types_in_public_api
 class DiceCounter = _DiceCounter with _$DiceCounter;
 
 // The store-class