+++ mobx_examples/lib/form/form_widgets.dart
@@ -45,9 +45,9 @@ class _FormExampleState extends State<FormExample> {
               ),
               Observer(
                   builder: (_) => AnimatedOpacity(
-                      child: const LinearProgressIndicator(),
                       duration: const Duration(milliseconds: 300),
-                      opacity: store.isUserCheckPending ? 1 : 0)),
+                      opacity: store.isUserCheckPending ? 1 : 0,
+                      child: const LinearProgressIndicator())),
               Observer(
                 builder: (_) => TextField(
                   onChanged: (value) => store.email = value,
@@ -67,8 +67,8 @@ class _FormExampleState extends State<FormExample> {
                 ),
               ),
               ElevatedButton(
-                child: const Text('Sign up'),
                 onPressed: store.validateAll,
+                child: const Text('Sign up'),
               )
             ],
           ),