+++ mobx_examples/lib/github/github_store.dart
@@ -3,6 +3,7 @@ import 'package:mobx/mobx.dart';
 
 part 'github_store.g.dart';
 
+// ignore: library_private_types_in_public_api
 class GithubStore = _GithubStore with _$GithubStore;
 
 abstract class _GithubStore with Store {