+++ tool/coverage.sh
@@ -17,16 +17,16 @@ dart \
 
 # Run the coverage collector to generate the JSON coverage report.
 echo "Collecting coverage..."
-nohup pub run coverage:collect_coverage \
+nohup dart run coverage:collect_coverage \
   --port=$OBS_PORT \
   --out=coverage/coverage.json \
   --wait-paused \
   --resume-isolates
 
 echo "Generating LCOV report..."
-pub run coverage:format_coverage \
+dart run coverage:format_coverage \
   --lcov \
   --in=coverage/coverage.json \
   --out=coverage/lcov.info \
-  --packages=.packages \
+  --packages=.dart_tool/package_config.json \
   --report-on=lib